+++ app/client/components/MostPopularCourse.tsx
@@ -1,39 +1,39 @@
-"use client";
-import { useState, useEffect } from "react";
-import { courses } from '../data/most-popular-course';
-import CourseCard from './CourseCard';
-import { ChevronLeftIcon, ChevronRightIcon } from "@heroicons/react/16/solid";
+'use client'
+import { useState, useEffect } from 'react'
+import { courses } from '@/app/client/data/most-popular-course'
+import CourseCard from './CourseCard'
+import { ChevronLeftIcon, ChevronRightIcon } from '@heroicons/react/16/solid'
 
 const useResponsiveCards = () => {
-  const [cardsToShow, setCardsToShow] = useState<number>(3);
-  
+  const [cardsToShow, setCardsToShow] = useState<number>(3)
+
   useEffect(() => {
     const updateCards = () => {
-      if (window.innerWidth < 640) setCardsToShow(1);
-      else if (window.innerWidth < 1024) setCardsToShow(2);
-      else setCardsToShow(3);
-    };
-    
-    window.addEventListener('resize', updateCards);
-    updateCards();
-    return () => window.removeEventListener('resize', updateCards);
-  }, []);
-  
-  return cardsToShow;
-};
+      if (window.innerWidth < 640) setCardsToShow(1)
+      else if (window.innerWidth < 1024) setCardsToShow(2)
+      else setCardsToShow(3)
+    }
 
-export default function NewCourseCard() {
-  const [startIndex, setStartIndex] = useState<number>(0);
-  const cardsToShow = useResponsiveCards();
-  const maxIndex = Math.ceil(courses.length / cardsToShow) - 1;
+    window.addEventListener('resize', updateCards)
+    updateCards()
+    return () => window.removeEventListener('resize', updateCards)
+  }, [])
+
+  return cardsToShow
+}
+
+export default function MostPopularCourse() {
+  const [startIndex, setStartIndex] = useState<number>(0)
+  const cardsToShow = useResponsiveCards()
+  const maxIndex = Math.ceil(courses.length / cardsToShow) - 1
 
   const nextSlide = () => {
-    setStartIndex((prev) => Math.min(prev + 1, maxIndex));
-  };
+    setStartIndex((prev) => Math.min(prev + 1, maxIndex))
+  }
 
   const prevSlide = () => {
-    setStartIndex((prev) => Math.max(prev - 1, 0));
-  };
+    setStartIndex((prev) => Math.max(prev - 1, 0))
+  }
 
   return (
     <div className="flex flex-col justify-start py-12">
@@ -42,9 +42,11 @@ export default function NewCourseCard() {
       </h1>
       <div className="relative flex">
         <div className="flex overflow-hidden py-6">
-          <div 
+          <div
             className="flex transition-transform duration-300 ease-in-out"
-            style={{ transform: `translateX(-${startIndex * (100 / cardsToShow)}%)` }}
+            style={{
+              transform: `translateX(-${startIndex * (100 / cardsToShow)}%)`,
+            }}
           >
             {courses.map((course, index) => (
               <div key={index} className="w-full pl-12">
@@ -53,7 +55,7 @@ export default function NewCourseCard() {
             ))}
           </div>
         </div>
-        
+
         {startIndex > 0 && (
           <button
             onClick={prevSlide}
@@ -62,7 +64,7 @@ export default function NewCourseCard() {
             <ChevronLeftIcon className="w-6 h-6" />
           </button>
         )}
-        
+
         {startIndex + 1 < maxIndex && (
           <button
             onClick={nextSlide}
@@ -73,5 +75,5 @@ export default function NewCourseCard() {
         )}
       </div>
     </div>
-  );
-};
+  )
+}
