+++ app/client/components/home/CurvePage.tsx
