+++ app/client/components/video/course-detail/FAQPage.tsx
