+++ app/client/components/video/course-detail/MeetChefCurvePage.tsx
