+++ app/client/group/page.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import NavbarAfter from '../components/NavbarAfter'
-import CalendarPage from '../components/CalendarPage'
-import AdvertPage from '../components/AdvertPage'
-import UpcomingWorkshopCard from '../components/UpcomingWorkshopCard'
+import CalendarPage from '../components/group/CalendarPage'
+import AdvertPage from '../components/group/AdvertPage'
+import UpcomingWorkshopCard from '../components/group/UpcomingWorkshopCard'
 import Footer from '../components/Footer'
 
 export default function GroupPage() {
