+++ app/client/login/forget-password/page.tsx
@@ -1,10 +1,10 @@
-import ForgetPassword from '../../components/ForgetPassword'
+import ForgetPassword from '../../components/login&forget-password/ForgetPassword'
 import React from 'react'
 
 export default function Forgetpassword() {
   return (
-    <>
+    <div className="animate-fadeIn">
       <ForgetPassword />
-    </>
+    </div>
   )
 }
