+++ app/client/sign-up/page.tsx
@@ -1,4 +1,4 @@
-import SignUp from '../components/SignUp';
+import SignUp from '../components/sign-up/SignUp';
 
 export default function Signup() {
   return (
