+++ app/client/video/course-detail/[course_id]/page.tsx
@@ -1,9 +1,9 @@
 import NavbarAfter from '@/app/client/components/NavbarAfter'
-import CourseDetailFirstPage from '@/app/client/components/CourseDetailFirstPage'
-import MeetChefCurvePage from '@/app/client/components/MeetChefCurvePage'
-import IngredientPage from '@/app/client/components/IngredientPage'
-import JoinOurCoursePage from '@/app/client/components/JoinOurCoursePage'
-import FAQPage from '@/app/client/components/FAQPage'
+import CourseDetailFirstPage from '@/app/client/components/video/course-detail/CourseDetailFirstPage'
+import MeetChefCurvePage from '@/app/client/components/video/course-detail/MeetChefCurvePage'
+import IngredientPage from '@/app/client/components/video/course-detail/IngredientPage'
+import JoinOurCoursePage from '@/app/client/components/video/course-detail/JoinOurCoursePage'
+import FAQPage from '@/app/client/components/video/course-detail/FAQPage'
 import Footer from '@/app/client/components/Footer'
 
 export default function CourseDetailPage() {
