+++ app/client/video/page.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
 import NavbarAfter from '../components/NavbarAfter'
-import LearningPage from '../components/LearningPage'
-import FoodTypeCard from '../components/FoodTypeCard'
+import LearningPage from '../components/video/search/LearningPage'
+import FoodTypeCard from '../components/video/search/FoodTypeCard'
 import MostPopularCourse from '../components/MostPopularCourse'
-import NewCourseCard from '../components/NewCourseCard'
-import TopRateCourse from '../components/TopRateCourse'
-import OrderIngredients from '../components/OrderIngredientPage'
-import PersonalCourseCard from '../components/PersonalCourseCard'
+import NewCourseCard from '../components/video/search/NewCourseCard'
+import TopRateCourse from '../components/video/search/TopRateCourse'
+import OrderIngredients from '../components/video/search/OrderIngredientPage'
+import PersonalCourseCard from '../components/video/search/PersonalCourseCard'
 import ContactUs from '../components/ContactUs'
 import Footer from '../components/Footer'
