+++ app/page.tsx
@@ -1,9 +1,9 @@
 import NavbarBefore from './client/components/NavbarBefore'
-import LandingPage from './client/components/LandingPage'
-import Card from './client/components/MostPopularCourse'
-import CurvePage from './client/components/CurvePage'
+import LandingPage from './client/components/home/LandingPage'
+import MostPopularCourse from './client/components/MostPopularCourse'
+import CurvePage from './client/components/home/CurvePage'
 import RecentCourses from './client/components/RecentCourses'
-import RecommendedForyou from './client/components/RecommendedForyou'
+import RecommendedForYou from './client/components/RecommendedForYou'
 import ClientReview from './client/components/ClientReview'
 import ContactUs from './client/components/ContactUs'
 import Footer from './client/components/Footer'
@@ -13,10 +13,10 @@ export default function Home() {
     <div className="bg-gradient-to-b from-[#F0725C] to-[#FE3511] animate-fadeIn">
       <NavbarBefore />
       <LandingPage />
-      <Card />
+      <MostPopularCourse />
       <CurvePage />
       <RecentCourses />
-      <RecommendedForyou />
+      <RecommendedForYou />
       <ClientReview />
       <ContactUs />
       <Footer />
